Refactor: Rename Appwrite function directory and update paths
The Appwrite function responsible for creating user documents was previously located in the `functions/` directory. This commit moves it to a more specific `functions/create_user/` directory.

The `pubspec.yaml` and `main.dart` files have been moved to this new location. There are no functional changes to the code itself.